feat/refactor: update token infrastructure classes
