+++ kub-super-admin-backend/src/main/java/education/kub/superadmin/infrastructure/token/store/service/TokenStoreService.java
@@ -1,4 +1,4 @@
-package education.kub.superadmin.services.inter;
+package education.kub.superadmin.infrastructure.token.store.service;
 
 public interface TokenStoreService {
     void deleteAllSessions(Long userId);
